backend/requirements.txt
fastapi==0.79.0
pydantic==1.9.1
uvicorn==0.18.2
psycopg2-binary==2.8.6
